Ref: Move EchoSelectorScreen -> ui
